requirements.txt
pandas
numpy
yfinance
statsmodels
arch
newsapi-python
nltk
streamlit
click
python-dotenv